Split out entity base classes and helper functions.
