+++ custom_components/isy994/__init__.py
@@ -5,32 +5,15 @@ from typing import Optional
 from urllib.parse import urlparse
 
 from pyisy import ISY
-from pyisy.constants import (
-    COMMAND_FRIENDLY_NAME,
-    EVENT_PROPS_IGNORED,
-    ISY_VALUE_UNKNOWN,
-    PROTO_GROUP,
-    PROTO_INSTEON,
-    PROTO_PROGRAM,
-    PROTO_ZWAVE,
-)
-from pyisy.helpers import NodeProperty
 import voluptuous as vol
 
 from homeassistant import config_entries
-from homeassistant.components.binary_sensor import DOMAIN as PLATFORM_BINARY_SENSOR
-from homeassistant.components.climate.const import DOMAIN as PLATFORM_CLIMATE
-from homeassistant.components.fan import DOMAIN as PLATFORM_FAN
-from homeassistant.components.light import DOMAIN as PLATFORM_LIGHT
-from homeassistant.components.sensor import DOMAIN as PLATFORM_SENSOR
-from homeassistant.components.switch import DOMAIN as PLATFORM_SWITCH
-from homeassistant.const import CONF_HOST, CONF_PASSWORD, CONF_USERNAME, STATE_UNKNOWN
+from homeassistant.const import CONF_HOST, CONF_PASSWORD, CONF_USERNAME
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers import config_validation as cv
 import homeassistant.helpers.device_registry as dr
-from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.entity_registry import async_get_registry
-from homeassistant.helpers.typing import ConfigType, Dict
+from homeassistant.helpers.typing import ConfigType
 
 from .const import (
     _LOGGER,
@@ -46,17 +29,12 @@ from .const import (
     ISY994_NODES,
     ISY994_PROGRAMS,
     ISY994_VARIABLES,
-    ISY_GROUP_PLATFORM,
-    KEY_ACTIONS,
-    KEY_FOLDER,
-    KEY_MY_PROGRAMS,
-    KEY_STATUS,
     MANUFACTURER,
-    NODE_FILTERS,
     SUPPORTED_PLATFORMS,
     SUPPORTED_PROGRAM_PLATFORMS,
     UNDO_UPDATE_LISTENER,
 )
+from .helpers import _categorize_nodes, _categorize_programs, _categorize_variables
 from .services import async_setup_services, async_unload_services
 
 CONFIG_SCHEMA = vol.Schema(
@@ -81,291 +59,6 @@ CONFIG_SCHEMA = vol.Schema(
     },
     extra=vol.ALLOW_EXTRA,
 )
-
-
-def _check_for_node_def(hass_isy_data: dict, node, single_platform: str = None) -> bool:
-    """Check if the node matches the node_def_id for any platforms.
-
-    This is only present on the 5.0 ISY firmware, and is the most reliable
-    way to determine a device's type.
-    """
-    if not hasattr(node, "node_def_id") or node.node_def_id is None:
-        # Node doesn't have a node_def (pre 5.0 firmware most likely)
-        return False
-
-    node_def_id = node.node_def_id
-
-    platforms = SUPPORTED_PLATFORMS if not single_platform else [single_platform]
-    for platform in platforms:
-        if node_def_id in NODE_FILTERS[platform]["node_def_id"]:
-            hass_isy_data[ISY994_NODES][platform].append(node)
-            return True
-
-    return False
-
-
-def _check_for_insteon_type(
-    hass_isy_data: dict, node, single_platform: str = None
-) -> bool:
-    """Check if the node matches the Insteon type for any platforms.
-
-    This is for (presumably) every version of the ISY firmware, but only
-    works for Insteon device. "Node Server" (v5+) and Z-Wave and others will
-    not have a type.
-    """
-    if not hasattr(node, "protocol") or node.protocol != PROTO_INSTEON:
-        return False
-    if not hasattr(node, "type") or node.type is None:
-        # Node doesn't have a type (non-Insteon device most likely)
-        return False
-
-    device_type = node.type
-    platforms = SUPPORTED_PLATFORMS if not single_platform else [single_platform]
-    for platform in platforms:
-        if any(
-            [
-                device_type.startswith(t)
-                for t in set(NODE_FILTERS[platform]["insteon_type"])
-            ]
-        ):
-
-            # Hacky special-cases for certain devices with different platforms
-            # included as subnodes. Note that special-cases are not necessary
-            # on ISY 5.x firmware as it uses the superior NodeDefs method
-
-            # FanLinc, which has a light module as one of its nodes.
-            if platform == PLATFORM_FAN and str(node.address[-1]) in ["1"]:
-                hass_isy_data[ISY994_NODES][PLATFORM_LIGHT].append(node)
-                return True
-
-            # Thermostats, which has a "Heat" and "Cool" sub-node on address 2 and 3
-            if platform == PLATFORM_CLIMATE and str(node.address[-1]) in ["2", "3"]:
-                hass_isy_data[ISY994_NODES][PLATFORM_BINARY_SENSOR].append(node)
-                return True
-
-            # IOLincs which have a sensor and relay on 2 different nodes
-            if (
-                platform == PLATFORM_BINARY_SENSOR
-                and device_type.startswith("7.")
-                and str(node.address[-1]) in ["2"]
-            ):
-                hass_isy_data[ISY994_NODES][PLATFORM_SWITCH].append(node)
-
-            # Smartenit EZIO2X4
-            if (
-                platform == PLATFORM_SWITCH
-                and device_type.startswith("7.3.255.")
-                and str(node.address[-1]) in ["9", "A", "B", "C"]
-            ):
-                hass_isy_data[ISY994_NODES][PLATFORM_BINARY_SENSOR].append(node)
-
-            hass_isy_data[ISY994_NODES][platform].append(node)
-            return True
-
-    return False
-
-
-def _check_for_zwave_cat(
-    hass_isy_data: dict, node, single_platform: str = None
-) -> bool:
-    """Check if the node matches the ISY Z-Wave Category for any platforms.
-
-    This is for (presumably) every version of the ISY firmware, but only
-    works for Z-Wave Devices with the devtype.cat property.
-    """
-    if not hasattr(node, "protocol") or node.protocol != PROTO_ZWAVE:
-        return False
-
-    if not hasattr(node, "zwave_props") or node.zwave_props is None:
-        # Node doesn't have a device type category (non-Z-Wave device)
-        return False
-
-    device_type = node.zwave_props.category
-    platforms = SUPPORTED_PLATFORMS if not single_platform else [single_platform]
-    for platform in platforms:
-        if any(
-            [
-                device_type.startswith(t)
-                for t in set(NODE_FILTERS[platform]["zwave_cat"])
-            ]
-        ):
-
-            hass_isy_data[ISY994_NODES][platform].append(node)
-            return True
-
-    return False
-
-
-def _check_for_uom_id(
-    hass_isy_data: dict, node, single_platform: str = None, uom_list: list = None
-) -> bool:
-    """Check if a node's uom matches any of the platforms uom filter.
-
-    This is used for versions of the ISY firmware that report uoms as a single
-    ID. We can often infer what type of device it is by that ID.
-    """
-    if not hasattr(node, "uom") or node.uom is None:
-        # Node doesn't have a uom (Scenes for example)
-        return False
-
-    node_uom = set(map(str.lower, node.uom))
-
-    if uom_list:
-        if node_uom.intersection(uom_list):
-            hass_isy_data[ISY994_NODES][single_platform].append(node)
-            return True
-    else:
-        platforms = SUPPORTED_PLATFORMS if not single_platform else [single_platform]
-        for platform in platforms:
-            if node_uom.intersection(NODE_FILTERS[platform]["uom"]):
-                hass_isy_data[ISY994_NODES][platform].append(node)
-                return True
-
-    return False
-
-
-def _check_for_states_in_uom(
-    hass_isy_data: dict, node, single_platform: str = None, states_list: list = None
-) -> bool:
-    """Check if a list of uoms matches two possible filters.
-
-    This is for versions of the ISY firmware that report uoms as a list of all
-    possible "human readable" states. This filter passes if all of the possible
-    states fit inside the given filter.
-    """
-    if not hasattr(node, "uom") or node.uom is None:
-        # Node doesn't have a uom (Scenes for example)
-        return False
-
-    node_uom = set(map(str.lower, node.uom))
-
-    if states_list:
-        if node_uom == set(states_list):
-            hass_isy_data[ISY994_NODES][single_platform].append(node)
-            return True
-    else:
-        platforms = SUPPORTED_PLATFORMS if not single_platform else [single_platform]
-        for platform in platforms:
-            if node_uom == set(NODE_FILTERS[platform]["states"]):
-                hass_isy_data[ISY994_NODES][platform].append(node)
-                return True
-
-    return False
-
-
-def _is_sensor_a_binary_sensor(hass_isy_data: dict, node) -> bool:
-    """Determine if the given sensor node should be a binary_sensor."""
-    if _check_for_node_def(hass_isy_data, node, single_platform=PLATFORM_BINARY_SENSOR):
-        return True
-    if _check_for_insteon_type(
-        hass_isy_data, node, single_platform=PLATFORM_BINARY_SENSOR
-    ):
-        return True
-
-    # For the next two checks, we're providing our own set of uoms that
-    # represent on/off devices. This is because we can only depend on these
-    # checks in the context of already knowing that this is definitely a
-    # sensor device.
-    if _check_for_uom_id(
-        hass_isy_data,
-        node,
-        single_platform=PLATFORM_BINARY_SENSOR,
-        uom_list=["2", "78"],
-    ):
-        return True
-    if _check_for_states_in_uom(
-        hass_isy_data,
-        node,
-        single_platform=PLATFORM_BINARY_SENSOR,
-        states_list=["on", "off"],
-    ):
-        return True
-
-    return False
-
-
-def _categorize_nodes(
-    hass_isy_data: dict, nodes, ignore_identifier: str, sensor_identifier: str
-) -> None:
-    """Sort the nodes to their proper platforms."""
-    for (path, node) in nodes:
-        ignored = ignore_identifier in path or ignore_identifier in node.name
-        if ignored:
-            # Don't import this node as a device at all
-            continue
-
-        if hasattr(node, "protocol") and node.protocol == PROTO_GROUP:
-            hass_isy_data[ISY994_NODES][ISY_GROUP_PLATFORM].append(node)
-            continue
-
-        if sensor_identifier in path or sensor_identifier in node.name:
-            # User has specified to treat this as a sensor. First we need to
-            # determine if it should be a binary_sensor.
-            if _is_sensor_a_binary_sensor(hass_isy_data, node):
-                continue
-            hass_isy_data[ISY994_NODES][PLATFORM_SENSOR].append(node)
-            continue
-
-        # We have a bunch of different methods for determining the device type,
-        # each of which works with different ISY firmware versions or device
-        # family. The order here is important, from most reliable to least.
-        if _check_for_node_def(hass_isy_data, node):
-            continue
-        if _check_for_insteon_type(hass_isy_data, node):
-            continue
-        if _check_for_zwave_cat(hass_isy_data, node):
-            continue
-        if _check_for_uom_id(hass_isy_data, node):
-            continue
-        if _check_for_states_in_uom(hass_isy_data, node):
-            continue
-
-
-def _categorize_programs(hass_isy_data: dict, programs: dict) -> None:
-    """Categorize the ISY994 programs."""
-    for platform in SUPPORTED_PROGRAM_PLATFORMS:
-        try:
-            folder = programs[KEY_MY_PROGRAMS][f"HA.{platform}"]
-        except KeyError:
-            pass
-        else:
-            for dtype, _, node_id in folder.children:
-                if dtype != KEY_FOLDER:
-                    continue
-                entity_folder = folder[node_id]
-                try:
-                    status = entity_folder[KEY_STATUS]
-                    assert status.dtype == PROTO_PROGRAM, "Not a program"
-                    if platform != PLATFORM_BINARY_SENSOR:
-                        actions = entity_folder[KEY_ACTIONS]
-                        assert actions.dtype == PROTO_PROGRAM, "Not a program"
-                    else:
-                        actions = None
-                except (AttributeError, KeyError, AssertionError):
-                    _LOGGER.warning(
-                        "Program entity '%s' not loaded due "
-                        "to invalid folder structure.",
-                        entity_folder.name,
-                    )
-                    continue
-
-                entity = (entity_folder.name, status, actions)
-                hass_isy_data[ISY994_PROGRAMS][platform].append(entity)
-
-
-def _categorize_variables(hass_isy_data: dict, variables, identifier: str) -> None:
-    """Gather the ISY994 Variables to be added as sensors."""
-    try:
-        var_to_add = [
-            (vtype, vname, vid)
-            for (vtype, vname, vid) in variables.children
-            if identifier in vname
-        ]
-    except KeyError as err:
-        _LOGGER.error("Error adding ISY Variables: %s", err)
-    else:
-        for vtype, vname, vid in var_to_add:
-            hass_isy_data[ISY994_VARIABLES].append((vname, variables[vtype][vid]))
 
 
 async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
@@ -418,7 +111,7 @@ async def async_setup_entry(
         hass_isy_data[ISY994_NODES][platform] = []
 
     hass_isy_data[ISY994_PROGRAMS] = {}
-    for platform in SUPPORTED_PLATFORMS:
+    for platform in SUPPORTED_PROGRAM_PLATFORMS:
         hass_isy_data[ISY994_PROGRAMS][platform] = []
 
     hass_isy_data[ISY994_VARIABLES] = []
@@ -599,200 +292,3 @@ async def migrate_old_unique_ids(hass, platform, devices):
             registry.async_update_entity(
                 old_entity_id_2, new_unique_id=device.unique_id
             )
-
-
-class ISYEntity(Entity):
-    """Representation of an ISY994 device."""
-
-    _name: str = None
-
-    def __init__(self, node) -> None:
-        """Initialize the insteon device."""
-        self._node = node
-        self._attrs = {}
-        self._change_handler = None
-        self._control_handler = None
-
-    async def async_added_to_hass(self) -> None:
-        """Subscribe to the node change events."""
-        self._change_handler = self._node.status_events.subscribe(self.on_update)
-
-        if hasattr(self._node, "control_events"):
-            self._control_handler = self._node.control_events.subscribe(self.on_control)
-
-    def on_update(self, event: object) -> None:
-        """Handle the update event from the ISY994 Node."""
-        self.schedule_update_ha_state()
-
-    def on_control(self, event: NodeProperty) -> None:
-        """Handle a control event from the ISY994 Node."""
-        event_data = {
-            "entity_id": self.entity_id,
-            "control": event.control,
-            "value": event.value,
-            "formatted": event.formatted,
-            "uom": event.uom,
-            "precision": event.prec,
-        }
-
-        if event.value is None or event.control not in EVENT_PROPS_IGNORED:
-            # New state attributes may be available, update the state.
-            self.schedule_update_ha_state()
-
-        self.hass.bus.fire("isy994_control", event_data)
-
-    @property
-    def device_info(self):
-        """Return the device_info of the device."""
-        if hasattr(self._node, "protocol") and self._node.protocol == PROTO_GROUP:
-            # not a device
-            return None
-        uuid = self._node.isy.configuration["uuid"]
-        device_info = {
-            "name": self.name,
-            "identifiers": {},
-            "model": "Unknown",
-            "manufacturer": "Unknown",
-            "via_device": (DOMAIN, uuid),
-        }
-        if hasattr(self._node, "address"):
-            device_info["name"] += f" ({self._node.address})"
-        if hasattr(self._node, "primary_node"):
-            device_info["identifiers"] = {(DOMAIN, f"{uuid}_{self._node.primary_node}")}
-        # ISYv5 Device Types
-        if hasattr(self._node, "node_def_id") and self._node.node_def_id is not None:
-            device_info["model"] = self._node.node_def_id
-            # Numerical Device Type
-            if hasattr(self._node, "type") and self._node.type is not None:
-                device_info["model"] += f" {self._node.type}"
-        if hasattr(self._node, "protocol"):
-            device_info["manufacturer"] = self._node.protocol
-            if self._node.protocol == PROTO_ZWAVE:
-                device_info[
-                    "manufacturer"
-                ] += f" mfr_id:{self._node.zwave_props.mfr_id}"
-                device_info["model"] += (
-                    f" Type:{self._node.zwave_props.mfr_id} "
-                    f"ProductTypeID:{self._node.zwave_props.prod_type_id} "
-                    f"ProductID:{self._node.zwave_props.product_id}"
-                )
-        # Note: sw_version is not exposed by the ISY for the individual devices.
-
-        return device_info
-
-    @property
-    def unique_id(self) -> str:
-        """Get the unique identifier of the device."""
-        if hasattr(self._node, "address"):
-            return f"{self._node.isy.configuration['uuid']}_{self._node.address}"
-        return None
-
-    @property
-    def old_unique_id(self) -> str:
-        """Get the old unique identifier of the device."""
-        if hasattr(self._node, "address"):
-            return self._node.address
-        return None
-
-    @property
-    def name(self) -> str:
-        """Get the name of the device."""
-        return self._name or str(self._node.name)
-
-    @property
-    def should_poll(self) -> bool:
-        """No polling required since we're using the subscription."""
-        return False
-
-    @property
-    def value(self) -> int:
-        """Get the current value of the device."""
-        return self._node.status
-
-    @property
-    def state(self):
-        """Return the state of the ISY device."""
-        if self.value == ISY_VALUE_UNKNOWN:
-            return STATE_UNKNOWN
-        return super().state
-
-
-class ISYNodeEntity(ISYEntity):
-    """Representation of a ISY Nodebase (Node/Group) entity."""
-
-    @property
-    def device_state_attributes(self) -> Dict:
-        """Get the state attributes for the device.
-
-        The 'aux_properties' in the pyisy Node class are combined with the
-        other attributes which have been picked up from the event stream and
-        the combined result are returned as the device state attributes.
-        """
-        attr = {}
-        if hasattr(self._node, "aux_properties"):
-            # Cast as list due to RuntimeError if a new property is added while running.
-            for name, value in list(self._node.aux_properties.items()):
-                attr_name = COMMAND_FRIENDLY_NAME.get(name, name)
-                attr[attr_name] = str(value.formatted).lower()
-
-        # If a Group/Scene, set a property if the entire scene is on/off
-        if hasattr(self._node, "group_all_on"):
-            # pylint: disable=protected-access
-            attr["group_all_on"] = "on" if self._node.group_all_on else "off"
-
-        self._attrs.update(attr)
-        return self._attrs
-
-    def send_node_command(self, command):
-        """Respond to an entity service command call."""
-        if not hasattr(self._node, command):
-            _LOGGER.error(
-                "Invalid Service Call %s for device %s.", command, self.entity_id
-            )
-            return
-        getattr(self._node, command)()
-
-    def send_raw_node_command(
-        self, command, value=None, unit_of_measurement=None, parameters=None
-    ):
-        """Respond to an entity service raw command call."""
-        if not hasattr(self._node, "send_cmd"):
-            _LOGGER.error(
-                "Invalid Service Call %s for device %s.", command, self.entity_id
-            )
-            return
-        self._node.send_cmd(command, value, unit_of_measurement, parameters)
-
-
-class ISYProgramEntity(ISYEntity):
-    """Representation of an ISY994 program base."""
-
-    def __init__(self, name: str, node, actions=None) -> None:
-        """Initialize the ISY994 program-based entity."""
-        super().__init__(node)
-        self._name = name
-        self._actions = actions
-
-    @property
-    def icon(self) -> str:
-        """Get the icon for programs."""
-        return "mdi:script-text-outline"  # Matches isy program icon
-
-    @property
-    def device_state_attributes(self) -> Dict:
-        """Get the state attributes for the device."""
-        attr = {}
-        if self._actions:
-            attr["actions_enabled"] = self._actions.enabled
-            attr["actions_last_finished"] = self._actions.last_finished
-            attr["actions_last_run"] = self._actions.last_run
-            attr["actions_last_update"] = self._actions.last_update
-            attr["ran_else"] = self._actions.ran_else
-            attr["ran_then"] = self._actions.ran_then
-            attr["run_at_startup"] = self._actions.run_at_startup
-            attr["running"] = self._actions.running
-        attr["status_enabled"] = self._node.enabled
-        attr["status_last_finished"] = self._node.last_finished
-        attr["status_last_run"] = self._node.last_run
-        attr["status_last_update"] = self._node.last_update
-        return attr
