+++ custom_components/isy994/climate.py
@@ -35,7 +35,7 @@ from homeassistant.const import (
 )
 from homeassistant.helpers.typing import HomeAssistantType
 
-from . import ISYNodeEntity, migrate_old_unique_ids
+from . import migrate_old_unique_ids
 from .const import (
     _LOGGER,
     DOMAIN as ISY994_DOMAIN,
@@ -45,6 +45,7 @@ from .const import (
     ISY_HVAC_MODES,
     UOM_TO_STATES,
 )
+from .entity import ISYNodeEntity
 from .services import async_setup_device_services
 
 ISY_SUPPORTED_FEATURES = (
