+++ custom_components/isy994/const.py
@@ -94,13 +94,41 @@ DEFAULT_VAR_SENSOR_STRING = "HA."
 DEFAULT_RESTORE_LIGHT_STATE = False
 DEFAULT_TLS_VERSION = 1.1
 
-DEFAULT_ON_VALUE = 1
-DEFAULT_OFF_VALUE = 0
-
 KEY_ACTIONS = "actions"
 KEY_FOLDER = "folder"
 KEY_MY_PROGRAMS = "My Programs"
 KEY_STATUS = "status"
+
+SUPPORTED_PLATFORMS = [
+    PLATFORM_BINARY_SENSOR,
+    PLATFORM_SENSOR,
+    PLATFORM_LOCK,
+    PLATFORM_FAN,
+    PLATFORM_COVER,
+    PLATFORM_LIGHT,
+    PLATFORM_SWITCH,
+    PLATFORM_CLIMATE,
+]
+SUPPORTED_PROGRAM_PLATFORMS = [
+    PLATFORM_BINARY_SENSOR,
+    PLATFORM_LOCK,
+    PLATFORM_FAN,
+    PLATFORM_COVER,
+    PLATFORM_SWITCH,
+]
+
+SUPPORTED_BIN_SENS_CLASSES = ["moisture", "opening", "motion", "climate"]
+
+# ISY Scenes are more like Switches than Home Assistant Scenes
+# (they can turn off, and report their state)
+ISY_GROUP_PLATFORM = PLATFORM_SWITCH
+
+ISY994_ISY = "isy"
+ISY994_NODES = "isy994_nodes"
+ISY994_PROGRAMS = "isy994_programs"
+ISY994_VARIABLES = "isy994_variables"
+
+UNDO_UPDATE_LISTENER = "undo_update_listener"
 
 # Do not use the Home Assistant consts for the states here - we're matching exact API
 # responses, not using them for Home Assistant states
@@ -211,43 +239,6 @@ NODE_FILTERS = {
         "zwave_cat": ["140"],
     },
 }
-
-SUPPORTED_PLATFORMS = [
-    PLATFORM_BINARY_SENSOR,
-    PLATFORM_SENSOR,
-    PLATFORM_LOCK,
-    PLATFORM_FAN,
-    PLATFORM_COVER,
-    PLATFORM_LIGHT,
-    PLATFORM_SWITCH,
-    PLATFORM_CLIMATE,
-]
-SUPPORTED_PROGRAM_PLATFORMS = [
-    PLATFORM_BINARY_SENSOR,
-    PLATFORM_LOCK,
-    PLATFORM_FAN,
-    PLATFORM_COVER,
-    PLATFORM_SWITCH,
-]
-# ISY Scenes are more like Switches than Home Assistant Scenes
-# (they can turn off, and report their state)
-ISY_GROUP_PLATFORM = PLATFORM_SWITCH
-
-ISY994_ISY = "isy"
-ISY994_NODES = "isy994_nodes"
-ISY994_PROGRAMS = "isy994_programs"
-ISY994_VARIABLES = "isy994_variables"
-
-UNDO_UPDATE_LISTENER = "undo_update_listener"
-
-ISY_HVAC_MODES = [
-    HVAC_MODE_OFF,
-    HVAC_MODE_HEAT,
-    HVAC_MODE_COOL,
-    HVAC_MODE_HEAT_COOL,
-    HVAC_MODE_AUTO,
-    HVAC_MODE_FAN_ONLY,
-]
 
 UOM_FRIENDLY_NAME = {
     "1": "A",
@@ -509,6 +500,15 @@ UOM_TO_STATES = {
     "99": {7: FAN_ON, 8: FAN_AUTO},  # Insteon Thermostat Fan Mode
 }
 
+ISY_HVAC_MODES = [
+    HVAC_MODE_OFF,
+    HVAC_MODE_HEAT,
+    HVAC_MODE_COOL,
+    HVAC_MODE_HEAT_COOL,
+    HVAC_MODE_AUTO,
+    HVAC_MODE_FAN_ONLY,
+]
+
 HA_HVAC_TO_ISY = {
     HVAC_MODE_OFF: "off",
     HVAC_MODE_HEAT: "heat",
@@ -519,8 +519,6 @@ HA_HVAC_TO_ISY = {
 }
 
 HA_FAN_TO_ISY = {FAN_ON: "on", FAN_AUTO: "auto"}
-
-SUPPORTED_BIN_SENS_CLASSES = ["moisture", "opening", "motion", "climate"]
 
 ISY_BIN_SENS_DEVICE_TYPES = {
     "moisture": ["16.8.", "16.13.", "16.14."],
