+++ custom_components/isy994/cover.py
@@ -8,7 +8,7 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import STATE_CLOSED, STATE_OPEN, STATE_UNKNOWN
 from homeassistant.helpers.typing import HomeAssistantType
 
-from . import ISYNodeEntity, ISYProgramEntity, migrate_old_unique_ids
+from . import migrate_old_unique_ids
 from .const import (
     _LOGGER,
     DOMAIN as ISY994_DOMAIN,
@@ -16,6 +16,7 @@ from .const import (
     ISY994_PROGRAMS,
     UOM_TO_STATES,
 )
+from .entity import ISYNodeEntity, ISYProgramEntity
 from .services import async_setup_device_services
 
 
@@ -88,3 +89,8 @@ class ISYCoverProgram(ISYProgramEntity, CoverDevice):
         """Send the close cover command to the ISY994 cover program."""
         if not self._actions.run_else():
             _LOGGER.error("Unable to close the cover")
+
+    @property
+    def icon(self) -> str:
+        """Get the icon for programs."""
+        return None
