+++ custom_components/isy994/light.py
@@ -13,7 +13,7 @@ from homeassistant.const import STATE_UNKNOWN
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import HomeAssistantType
 
-from . import ISYNodeEntity, migrate_old_unique_ids
+from . import migrate_old_unique_ids
 from .const import (
     _LOGGER,
     ATTR_LAST_BRIGHTNESS,
@@ -21,6 +21,7 @@ from .const import (
     DOMAIN as ISY994_DOMAIN,
     ISY994_NODES,
 )
+from .entity import ISYNodeEntity
 from .services import async_setup_device_services, async_setup_light_services
