+++ custom_components/isy994/sensor.py
@@ -8,7 +8,7 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import STATE_UNKNOWN, TEMP_CELSIUS, TEMP_FAHRENHEIT
 from homeassistant.helpers.typing import HomeAssistantType
 
-from . import ISYEntity, ISYNodeEntity, migrate_old_unique_ids
+from . import migrate_old_unique_ids
 from .const import (
     _LOGGER,
     DOMAIN as ISY994_DOMAIN,
@@ -17,6 +17,7 @@ from .const import (
     UOM_FRIENDLY_NAME,
     UOM_TO_STATES,
 )
+from .entity import ISYEntity, ISYNodeEntity
 from .services import async_setup_device_services
